% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{cioddsr}
\alias{cioddsr}
\title{Odds Ratio and CI function}
\usage{
cioddsr(ecases, enoncases, uecases, uenoncases)
}
\arguments{
\item{ecases}{exposure cases}

\item{enoncases}{exposure noncases}

\item{uecases}{non exposed cases}

\item{uenoncases}{non exposed non cases}
}
\description{
This function calculates an adds ratio and the CI
}
\examples{
cioddsr_function()
}
\keyword{cioddsr}
