% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{cirater}
\alias{cirater}
\title{Rate Ratio and CI function}
\usage{
cirater(cases, noncases, pyrcases, pyrnoncases)
}
\arguments{
\item{cases}{cases}

\item{noncases}{noncases}

\item{pyrcases}{person years for cases}

\item{pyrnoncases}{person years for noncases}
}
\description{
This function calculates a Rate ratio and the CI
}
\examples{
cirater_function()
}
\keyword{cirater}
