% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{ciriskr}
\alias{ciriskr}
\title{Risk Ratio and CI function}
\usage{
ciriskr(ecases, enoncases, uecases, uenoncases)
}
\arguments{
\item{ecases}{exposure cases}

\item{enoncases}{exposure noncases}

\item{uecases}{non exposed cases}

\item{uenoncases}{non exposed non cases}
}
\description{
This function calculates a risk ratio and the CI
}
\examples{
ciriskr_function()
}
\keyword{ciriskr}
