% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{etable}
\alias{etable}
\title{2x2 Epi Table Function}
\usage{
etable(exposure, disease)
}
\arguments{
\item{exposure}{exposure column}

\item{disease}{disease column}
}
\description{
This function creates a 2x2 Disease Exposure Table
}
\examples{
etable_function()
}
\keyword{etable}
