% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{etable2}
\alias{etable2}
\title{2x2 Epi Table Function with margins}
\usage{
etable2(exposure, disease)
}
\arguments{
\item{exposure}{exposure column}

\item{disease}{disease column}
}
\description{
This function creates a 2x2 Disease Exposure Table with margins
}
\examples{
etable2_function()
}
\keyword{etable2}
