% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{rateratio}
\alias{rateratio}
\title{Rate Ratio}
\usage{
rateratio(cases, noncases, pyrcases, pyrnoncases)
}
\arguments{
\item{cases}{cases}

\item{noncases}{noncases}

\item{pyrcases}{person years for cases}

\item{pyrnoncases}{person years for noncases}
}
\description{
This function calculates a rate ratio
}
\examples{
rateratio_function()
}
\keyword{rateratio}
