% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{riskratio}
\alias{riskratio}
\title{Risk Ratio function}
\usage{
riskratio(a, b, c, d)
}
\arguments{
\item{a}{first entry}

\item{b}{second entry}

\item{c}{third entry}

\item{d}{fourth entry}
}
\description{
This function calculates a risk ratio
}
\examples{
riskratio_function()
}
\keyword{risk}
